Update max sp config
Specify a NameId format that MAX.gov supports. Ensure that `maxEmail` is
provided, since we depend on that for identifying the user. Since all the
attributes provided by MAX are custom, we need to allow_unknown_attributes to
avoid validation errors. I'm not sure if/how they could be provided in the attribute maps.